requirements.txt
# AI Blockchain - Zależności Pythona
# Wersja: 1.0.0
# Data: 2025-12-09

# === Podstawowe ===
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2
tqdm>=4.66.0  # Progress bars

# === Przetwarzanie Danych ===
pandas>=1.5.0,<2.1.0  # Kompatybilne z Python 3.9
numpy>=1.24.0,<1.26.0 # Kompatybilne z Python 3.9
pyarrow>=10.0.0,<14.0.0

# === API Giełd ===
ccxt>=4.1.0           # Ujednolicone API dla 100+ giełd
python-binance>=1.0.19
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# === Analiza Techniczna ===
# pandas-ta - wymaga Python >=3.12, użyjemy alternatyw
ta>=0.11.0            # Biblioteka wskaźników technicznych
mplfinance>=0.12.9b7  # Wykresy świecowe (najnowsza dla Python 3.9)
# ta-lib>=0.4.28      # TA-Lib (opcjonalnie, wymaga instalacji C library)

# === Machine Learning ===
scikit-learn>=1.0.0,<1.3.0  # Kompatybilne z Python 3.9
torch>=1.13.0,<2.1.0         # Kompatybilne z Python 3.9 (opcjonalnie)
xgboost>=1.7.0,<2.0.0        # Opcjonalnie
# lightgbm>=3.3.0,<4.1.0    # Wymaga kompilacji, opcjonalnie

# === Deep Learning - Szeregi Czasowe ===
# pytorch-forecasting>=1.0.0  # Może wymagać nowszego Pythona
# darts>=0.27.0                # Może wymagać nowszego Pythona

# === LLM & NLP ===
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.6.0
anthropic>=0.8.0
tiktoken>=0.5.0

# Analiza sentymentu
transformers>=4.36.0
sentence-transformers>=2.2.0
vaderSentiment>=3.3.2
praw>=7.7.0              # Reddit API wrapper

# === Wizualizacje ===
matplotlib>=3.8.0
plotly>=5.18.0
dash>=2.14.0
seaborn>=0.13.0

# === API Backend ===
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0

# === Bazy Danych ===
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9      # PostgreSQL driver
redis>=5.0.0
alembic>=1.13.0             # Migracje bazy danych
python-dotenv>=1.0.0        # Ładowanie .env

# === TimescaleDB ===
# Instalacja rozszerzenia w PostgreSQL:
# CREATE EXTENSION IF NOT EXISTS timescaledb;

# === Scheduler & Tasks ===
apscheduler>=3.10.0
celery>=5.3.0

# === Testy ===
pytest>=7.4.0
pytest-asyncio>=0.23.0

# === Dev Tools ===
black>=23.12.0
ruff>=0.1.9
mypy>=1.8.0
ipykernel>=6.27.0
jupyter>=1.0.0

